requirements.txt
alembic                == 1.9.2
async-generator        == 1.10
attrs                  == 22.1.0
Automat                == 22.10.0
autopep8               == 2.0.1
bcrypt                 == 4.0.1
blinker                == 1.5
bulma                  == 1.0.1
cachetools             == 5.2.0
certifi                == 2022.9.24
cffi                   == 1.15.1
charset-normalizer     == 2.1.1
click                  == 8.1.3
constantly             == 15.1.0
cryptography           == 38.0.3
cssselect              == 1.2.0
dnspython              == 2.2.1
email-validator        == 1.3.0
exceptiongroup         == 1.0.4
fake-useragent         == 1.1.0
filelock               == 3.8.0
Flask                  == 2.2.2
Flask-Assets           == 2.0
Flask-Bcrypt           == 1.0.1
Flask-Login            == 0.6.2
Flask-Mail             == 0.9.1
Flask-Migrate          == 4.0.1
Flask-SQLAlchemy       == 3.0.2
Flask-WTF              == 1.0.1
google-auth            == 2.15.0
google-auth-oauthlib   == 0.8.0
greenlet               == 2.0.1
gspread                == 5.7.2
h11                    == 0.14.0
hyperlink              == 21.0.0
idna                   == 3.4
incremental            == 22.10.0
install                == 1.3.5
itemadapter            == 0.7.0
itemloaders            == 1.0.6
itsdangerous           == 2.1.2
Jinja2                 == 3.1.2
jmespath               == 1.0.1
lxml                   == 4.9.1
Mako                   == 1.2.4
MarkupSafe             == 2.1.1
oauth                  == 1.0.1
oauthlib               == 3.2.2
outcome                == 1.2.0
packaging              == 21.3
parsel                 == 1.7.0
Password-Validation    == 0.1.1
Protego                == 0.2.1
psycopg                == 3.1.4
psycopg-binary         == 3.1.4
psycopg-pool           == 3.1.4
psycopg2               == 2.9.5
pyasn1                 == 0.4.8
pyasn1-modules         == 0.2.8
pycodestyle            == 2.10.0
pycparser              == 2.21
PyDispatcher           == 2.0.6
pygame                 == 2.1.2
PyJWT                  == 2.6.0
pyOpenSSL              == 22.1.0
pyparsing              == 3.0.9
PyQt6                  == 6.4.0
PyQt6-Qt6              == 6.4.1
PyQt6-sip              == 13.4.0
PySocks                == 1.7.1
python-dotenv          == 0.21.0
pytz                   == 2022.7
queuelib               == 1.6.2
requests               == 2.28.1
requests-file          == 1.5.1
requests-oauthlib      == 1.3.1
rsa                    == 4.9
Scrapy                 == 2.7.1
selenium               == 4.6.1
selenium-stealth       == 1.0.6
service-identity       == 21.1.0
setuptools             == 63.2.0
six                    == 1.16.0
sniffio                == 1.3.0
sortedcontainers       == 2.4.0
SQLAlchemy             == 1.4.46
tldextract             == 3.4.0
tomli                  == 2.0.1
trio                   == 0.22.0
trio-websocket         == 0.9.2
twilio                 == 7.16.0
Twisted                == 22.10.0
typing_extensions      == 4.4.0
undetected-chromedriver== 3.1.7
urllib3                == 1.26.13
w3lib                  == 2.0.1
webassets              == 2.0
websockets             == 10.4
Werkzeug               == 2.2.2
wsproto                == 1.2.0
WTForms                == 3.0.1
zope.interface         == 5.5.2

